Allowed users to enter receiver's address on UI. Updated server.py and frontend to work exchange sender, receiver and message
