+++ frontend/src/components/ChatInterface/ChatBanner.jsx
@@ -1,0 +1,9 @@
+export default function ChatBanner({ receiver }) {
+  return (
+    <>
+      <div className="alert alert-secondary me-3">
+        <h1>{receiver}</h1>
+      </div>
+    </>
+  );
+}
